fixup! Minor refactor in pprof aggregation Extract the label creation logic into a function
